+++ data-model-db/src/main/java/org/mousephenotype/cda/db/pojo/CategoricalResult.java
@@ -282,11 +282,11 @@ public class CategoricalResult extends StatisticalResult implements Serializable
         return true;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return this.id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -299,6 +299,4 @@ public class CategoricalResult extends StatisticalResult implements Serializable
 		this.mpTermId=mpTermId;
 
 	}
-
-
 }
