+++ data-model-db/src/main/java/org/mousephenotype/cda/db/pojo/Organisation.java
@@ -36,7 +36,7 @@ public class Organisation implements Serializable {
 
 	@Id
 	@Column(name = "id")
-	private Integer id;
+	private Long id;
 
 	@Column(name = "name")
 	private String name;
@@ -69,14 +69,14 @@ public class Organisation implements Serializable {
 	/**
 	 * @return the id
 	 */
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
 	/**
 	 * @param id the id to set
 	 */
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -172,5 +172,4 @@ public class Organisation implements Serializable {
 				+ fullname + ", country=" + country + ", projects=" + projects
 				+ "]";
 	}
-
 }
